updt calc

+++ src/s01-mul.cpp
@@ -1,10 +1,11 @@
-#include <iostream>
+#include <iostream> 
 #include <string>
 
-auto main(int, char*argv[])->int
+auto main(int, char* argv[]) -> int
 {
-  auto const a=std::stoll(argv[1]);
-  auto const b=std::stoll(argv[2]);
-  std::cout<<(a*b)<<"\n";
-  return 0;
+   auto const a = std::stoi(argv[1]);
+   auto const b = std::stoi(argv[2]);
+   std::cout << (a * b) << "\n";
+   return 0;
 }
+
